Moved unit tests to CXX files to test C++ bindings implicitly.
