+++ unit/CMakeLists.txt
@@ -40,82 +40,82 @@ target_link_options(open1722 PRIVATE -coverage)
 target_compile_options(open1722custom PRIVATE -coverage)
 target_link_options(open1722custom PRIVATE -coverage)
 
-add_executable(test-can test-can.c)
+add_executable(test-can test-can.cpp)
 target_link_libraries(test-can open1722 cmocka m)
 target_include_directories(test-can PUBLIC ../include)
 add_test(NAME test-can COMMAND test-can)
 
-add_executable(test-aaf test-aaf.c)
+add_executable(test-aaf test-aaf.cpp)
 target_link_libraries(test-aaf open1722 cmocka)
 target_include_directories(test-aaf PUBLIC ../include)
 add_test(NAME test-aaf COMMAND test-aaf)
 
-add_executable(test-avtp test-avtp.c)
+add_executable(test-avtp test-avtp.cpp)
 target_link_libraries(test-avtp open1722 cmocka)
 target_include_directories(test-avtp PUBLIC ../include)
 add_test(NAME test-avtp COMMAND test-avtp)
 
-add_executable(test-crf test-crf.c)
+add_executable(test-crf test-crf.cpp)
 target_link_libraries(test-crf open1722 cmocka)
 target_include_directories(test-crf PUBLIC ../include)
 add_test(NAME test-crf COMMAND test-crf)
 
-add_executable(test-cvf test-cvf.c)
+add_executable(test-cvf test-cvf.cpp)
 target_link_libraries(test-cvf open1722 cmocka)
 target_include_directories(test-cvf PUBLIC ../include)
 add_test(NAME test-cvf COMMAND test-cvf)
 
-add_executable(test-rvf test-rvf.c)
+add_executable(test-rvf test-rvf.cpp)
 target_link_libraries(test-rvf open1722 cmocka)
 target_include_directories(test-rvf PUBLIC ../include)
 add_test(NAME test-rvf COMMAND test-rvf)
 
-add_executable(test-tscf test-tscf.c)
+add_executable(test-tscf test-tscf.cpp)
 target_link_libraries(test-tscf open1722 cmocka)
 target_include_directories(test-tscf PUBLIC ../include)
 add_test(NAME test-tscf COMMAND test-tscf)
 
-add_executable(test-ntscf test-ntscf.c)
+add_executable(test-ntscf test-ntscf.cpp)
 target_link_libraries(test-ntscf open1722 cmocka)
 target_include_directories(test-ntscf PUBLIC ../include)
 add_test(NAME test-ntscf COMMAND test-ntscf)
 
-add_executable(test-vss test-vss.c)
+add_executable(test-vss test-vss.cpp)
 target_link_libraries(test-vss open1722 open1722custom cmocka)
 target_include_directories(test-vss PUBLIC ../include)
 add_test(NAME test-vss COMMAND test-vss)
 
-add_executable(test-flexray test-flexray.c)
+add_executable(test-flexray test-flexray.cpp)
 target_link_libraries(test-flexray open1722 cmocka)
 target_include_directories(test-flexray PUBLIC ../include)
 add_test(NAME test-flexray COMMAND test-flexray)
 
-add_executable(test-lin test-lin.c)
+add_executable(test-lin test-lin.cpp)
 target_link_libraries(test-lin open1722 cmocka)
 target_include_directories(test-lin PUBLIC ../include)
 add_test(NAME test-lin COMMAND test-lin)
 
-add_executable(test-most test-most.c)
+add_executable(test-most test-most.cpp)
 target_link_libraries(test-most open1722 cmocka)
 target_include_directories(test-most PUBLIC ../include)
 add_test(NAME test-most COMMAND test-most)
 
-add_executable(test-gpc test-gpc.c)
+add_executable(test-gpc test-gpc.cpp)
 target_link_libraries(test-gpc open1722 cmocka)
 target_include_directories(test-gpc PUBLIC ../include)
 add_test(NAME test-gpc COMMAND test-gpc)
 
-add_executable(test-sensor test-sensor.c)
+add_executable(test-sensor test-sensor.cpp)
 target_link_libraries(test-sensor open1722 cmocka)
 target_include_directories(test-sensor PUBLIC ../include)
 add_test(NAME test-sensor COMMAND test-sensor)
 
-add_executable(test-udp test-udp.c)
+add_executable(test-udp test-udp.cpp)
 target_link_libraries(test-udp open1722 cmocka)
 target_include_directories(test-udp PUBLIC ../include)
 add_test(NAME test-udp COMMAND test-udp)
 
-add_executable(test-commonheader test-commonheader.c)
+add_executable(test-commonheader test-commonheader.cpp)
 target_link_libraries(test-commonheader open1722 cmocka)
 target_include_directories(test-commonheader PUBLIC ../include)
 add_test(NAME test-commonheader COMMAND test-commonheader)
