+++ unit/test-crf.cpp
@@ -28,7 +28,9 @@
 #include <stdarg.h>
 #include <stddef.h>
 #include <setjmp.h>
+extern "C" {
 #include <cmocka.h>
+}
 #include <arpa/inet.h>
 #include <stdio.h>
 #include <errno.h>
