+++ unit/test-cvf.cpp
@@ -1,5 +1,6 @@
 /*
- * Copyright (c) 2017, Intel Corporation
+ * Copyright (c) 2019, Intel Corporation
+ * Copyright (c) 2024, COVESA
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
@@ -25,48 +26,53 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
+#include <alloca.h>
 #include <stdarg.h>
 #include <stddef.h>
+#include <string.h>
 #include <setjmp.h>
+extern "C" {
 #include <cmocka.h>
+}
 #include <arpa/inet.h>
 #include <errno.h>
 
 #include "avtp/CommonHeader.h"
-#include "avtp/aaf/Pcm.h"
+#include "avtp/cvf/Cvf.h"
+#include "avtp/cvf/H264.h"
 
-static void aaf_get_field_null_pdu(void **state)
+static void cvf_get_field_null_pdu(void **state)
 {
     int res;
     uint64_t val = 1;
 
-    res = avtp_aaf_pdu_get(NULL, AVTP_AAF_FIELD_SV, &val);
+    res = avtp_cvf_pdu_get(NULL, AVTP_CVF_FIELD_SV, &val);
 
     assert_int_equal(res, -EINVAL);
 }
 
-static void aaf_get_field_null_val(void **state)
+static void cvf_get_field_null_val(void **state)
 {
     int res;
     struct avtp_stream_pdu pdu = { 0 };
 
-    res = avtp_aaf_pdu_get(&pdu, AVTP_AAF_FIELD_SV, NULL);
+    res = avtp_cvf_pdu_get(&pdu, AVTP_CVF_FIELD_SV, NULL);
 
     assert_int_equal(res, -EINVAL);
 }
 
-static void aaf_get_field_invalid_field(void **state)
+static void cvf_get_field_invalid_field(void **state)
 {
     int res;
     uint64_t val = 1;
     struct avtp_stream_pdu pdu = { 0 };
 
-    res = avtp_aaf_pdu_get(&pdu, AVTP_AAF_FIELD_MAX, &val);
+    res = avtp_cvf_pdu_get(&pdu, AVTP_CVF_FIELD_MAX, &val);
 
     assert_int_equal(res, -EINVAL);
 }
 
-static void aaf_get_field_sv(void **state)
+static void cvf_get_field_sv(void **state)
 {
     int res;
     uint64_t val;
@@ -75,13 +81,13 @@ static void aaf_get_field_sv(void **state)
     /* Set 'sv' field to 1. */
     pdu.subtype_data = htonl(0x00800000);
 
-    res = avtp_aaf_pdu_get(&pdu, AVTP_AAF_FIELD_SV, &val);
+    res = avtp_cvf_pdu_get(&pdu, AVTP_CVF_FIELD_SV, &val);
 
     assert_int_equal(res, 0);
     assert_true(val == 1);
 }
 
-static void aaf_get_field_mr(void **state)
+static void cvf_get_field_mr(void **state)
 {
     int res;
     uint64_t val;
@@ -90,13 +96,13 @@ static void aaf_get_field_mr(void **state)
     /* Set 'mr' field to 1. */
     pdu.subtype_data = htonl(0x00080000);
 
-    res = avtp_aaf_pdu_get(&pdu, AVTP_AAF_FIELD_MR, &val);
+    res = avtp_cvf_pdu_get(&pdu, AVTP_CVF_FIELD_MR, &val);
 
     assert_int_equal(res, 0);
     assert_true(val == 1);
 }
 
-static void aaf_get_field_tv(void **state)
+static void cvf_get_field_tv(void **state)
 {
     int res;
     uint64_t val;
@@ -105,13 +111,13 @@ static void aaf_get_field_tv(void **state)
     /* Set 'tv' field to 1. */
     pdu.subtype_data = htonl(0x00010000);
 
-    res = avtp_aaf_pdu_get(&pdu, AVTP_AAF_FIELD_TV, &val);
+    res = avtp_cvf_pdu_get(&pdu, AVTP_CVF_FIELD_TV, &val);
 
     assert_int_equal(res, 0);
     assert_true(val == 1);
 }
 
-static void aaf_get_field_seq_num(void **state)
+static void cvf_get_field_seq_num(void **state)
 {
     int res;
     uint64_t val;
@@ -120,13 +126,13 @@ static void aaf_get_field_seq_num(void **state)
     /* Set 'sequence_num' field to 0x55. */
     pdu.subtype_data = htonl(0x00005500);
 
-    res = avtp_aaf_pdu_get(&pdu, AVTP_AAF_FIELD_SEQ_NUM, &val);
+    res = avtp_cvf_pdu_get(&pdu, AVTP_CVF_FIELD_SEQUENCE_NUM, &val);
 
     assert_int_equal(res, 0);
     assert_true(val == 0x55);
 }
 
-static void aaf_get_field_tu(void **state)
+static void cvf_get_field_tu(void **state)
 {
     int res;
     uint64_t val;
@@ -135,13 +141,13 @@ static void aaf_get_field_tu(void **state)
     /* Set 'tu' field to 1. */
     pdu.subtype_data = htonl(0x00000001);
 
-    res = avtp_aaf_pdu_get(&pdu, AVTP_AAF_FIELD_TU, &val);
+    res = avtp_cvf_pdu_get(&pdu, AVTP_CVF_FIELD_TU, &val);
 
     assert_int_equal(res, 0);
     assert_true(val == 1);
 }
 
-static void aaf_get_field_stream_id(void **state)
+static void cvf_get_field_stream_id(void **state)
 {
     int res;
     uint64_t val;
@@ -150,13 +156,13 @@ static void aaf_get_field_stream_id(void **state)
     /* Set 'stream_id' field to 0xAABBCCDDEEFF0001. */
     pdu.stream_id = htobe64(0xAABBCCDDEEFF0001);
 
-    res = avtp_aaf_pdu_get(&pdu, AVTP_AAF_FIELD_STREAM_ID, &val);
+    res = avtp_cvf_pdu_get(&pdu, AVTP_CVF_FIELD_STREAM_ID, &val);
 
     assert_int_equal(res, 0);
     assert_true(val == 0xAABBCCDDEEFF0001);
 }
 
-static void aaf_get_field_timestamp(void **state)
+static void cvf_get_field_timestamp(void **state)
 {
     int res;
     uint64_t val;
@@ -165,73 +171,43 @@ static void aaf_get_field_timestamp(void **state)
     /* Set 'avtp_timestamp' field to 0x80C0FFEE. */
     pdu.avtp_time = htonl(0x80C0FFEE);
 
-    res = avtp_aaf_pdu_get(&pdu, AVTP_AAF_FIELD_TIMESTAMP, &val);
+    res = avtp_cvf_pdu_get(&pdu, AVTP_CVF_FIELD_AVTP_TIMESTAMP, &val);
 
     assert_int_equal(res, 0);
     assert_true(val == 0x80C0FFEE);
 }
 
-static void aaf_get_field_format(void **state)
+static void cvf_get_field_format(void **state)
 {
     int res;
     uint64_t val;
     struct avtp_stream_pdu pdu = { 0 };
 
-    /* Set 'format' field to AVTP_AAF_FORMAT_INT_16BIT. */
-    pdu.format_specific = htonl(0x04000000);
+    /* Set 'format' field to AVTP_CVF_FORMAT_RFC. */
+    pdu.format_specific = htonl(0x02000000);
 
-    res = avtp_aaf_pdu_get(&pdu, AVTP_AAF_FIELD_FORMAT, &val);
+    res = avtp_cvf_pdu_get(&pdu, AVTP_CVF_FIELD_FORMAT, &val);
 
     assert_int_equal(res, 0);
-    assert_true(val == AVTP_AAF_FORMAT_INT_16BIT);
+    assert_true(val == AVTP_CVF_FORMAT_RFC);
 }
 
-static void aaf_get_field_nsr(void **state)
+static void cvf_get_field_format_subtype(void **state)
 {
     int res;
     uint64_t val;
     struct avtp_stream_pdu pdu = { 0 };
 
-    /* Set 'nsr' field to AVTP_AAF_PCM_NSR_48KHZ. */
-    pdu.format_specific = htonl(0x00500000);
+    /* Set 'format_subtype' field to AVTP_CVF_FORMAT_SUBTYPE_H264. */
+    pdu.format_specific = htonl(0x00010000);
 
-    res = avtp_aaf_pdu_get(&pdu, AVTP_AAF_FIELD_NSR, &val);
-
-    assert_int_equal(res, 0);
-    assert_true(val == AVTP_AAF_PCM_NSR_48KHZ);
-}
-
-static void aaf_get_field_chan(void **state)
-{
-    int res;
-    uint64_t val;
-    struct avtp_stream_pdu pdu = { 0 };
-
-    /* Set 'channels_per_frame' field to 0x2AA. */
-    pdu.format_specific = htonl(0x0002AA00);
-
-    res = avtp_aaf_pdu_get(&pdu, AVTP_AAF_FIELD_CHAN_PER_FRAME, &val);
+    res = avtp_cvf_pdu_get(&pdu, AVTP_CVF_FIELD_FORMAT_SUBTYPE, &val);
 
     assert_int_equal(res, 0);
-    assert_true(val == 0x2AA);
+    assert_true(val == AVTP_CVF_FORMAT_SUBTYPE_H264);
 }
 
-static void aaf_get_field_depth(void **state)
-{
-    int res;
-    uint64_t val;
-    struct avtp_stream_pdu pdu = { 0 };
-
-    /* Set 'bit_depth' field to 0xA5. */
-    pdu.format_specific = htonl(0x000000A5);
-
-    res = avtp_aaf_pdu_get(&pdu, AVTP_AAF_FIELD_BIT_DEPTH, &val);
-
-    assert_int_equal(res, 0);
-    assert_true(val == 0xA5);
-}
-
-static void aaf_get_field_data_len(void **state)
+static void cvf_get_field_data_len(void **state)
 {
     int res;
     uint64_t val;
@@ -240,28 +216,28 @@ static void aaf_get_field_data_len(void **state)
     /* Set 'stream_data_length' field to 0xAAAA. */
     pdu.packet_info = htonl(0xAAAA0000);
 
-    res = avtp_aaf_pdu_get(&pdu, AVTP_AAF_FIELD_STREAM_DATA_LEN, &val);
+    res = avtp_cvf_pdu_get(&pdu, AVTP_CVF_FIELD_STREAM_DATA_LENGTH, &val);
 
     assert_int_equal(res, 0);
     assert_true(val == 0xAAAA);
 }
 
-static void aaf_get_field_sp(void **state)
+static void cvf_get_field_m(void **state)
 {
     int res;
     uint64_t val;
     struct avtp_stream_pdu pdu = { 0 };
 
-    /* Set 'sp' field to AVTP_AAF_PCM_SP_SPARSE. */
+    /* Set 'M' field to 0x1. */
     pdu.packet_info = htonl(0x00001000);
 
-    res = avtp_aaf_pdu_get(&pdu, AVTP_AAF_FIELD_SP, &val);
+    res = avtp_cvf_pdu_get(&pdu, AVTP_CVF_FIELD_M, &val);
 
     assert_int_equal(res, 0);
-    assert_true(val == AVTP_AAF_PCM_SP_SPARSE);
+    assert_true(val == 0x1);
 }
 
-static void aaf_get_field_evt(void **state)
+static void cvf_get_field_evt(void **state)
 {
     int res;
     uint64_t val;
@@ -270,37 +246,37 @@ static void aaf_get_field_evt(void **state)
     /* Set 'evt' field to 0xA. */
     pdu.packet_info = htonl(0x00000A00);
 
-    res = avtp_aaf_pdu_get(&pdu, AVTP_AAF_FIELD_EVT, &val);
+    res = avtp_cvf_pdu_get(&pdu, AVTP_CVF_FIELD_EVT, &val);
 
     assert_int_equal(res, 0);
     assert_true(val == 0xA);
 }
 
-static void aaf_set_field_null_pdu(void **state)
+static void cvf_set_field_null_pdu(void **state)
 {
     int res;
 
-    res = avtp_aaf_pdu_set(NULL, AVTP_AAF_FIELD_SV, 1);
+    res = avtp_cvf_pdu_set(NULL, AVTP_CVF_FIELD_SV, 1);
 
     assert_int_equal(res, -EINVAL);
 }
 
-static void aaf_set_field_invalid_field(void **state)
+static void cvf_set_field_invalid_field(void **state)
 {
     int res;
     struct avtp_stream_pdu pdu = { 0 };
 
-    res = avtp_aaf_pdu_set(&pdu, AVTP_AAF_FIELD_MAX, 1);
+    res = avtp_cvf_pdu_set(&pdu, AVTP_CVF_FIELD_MAX, 1);
 
     assert_int_equal(res, -EINVAL);
 }
 
-static void aaf_set_field_sv(void **state)
+static void cvf_set_field_sv(void **state)
 {
     int res;
     struct avtp_stream_pdu pdu = { 0 };
 
-    res = avtp_aaf_pdu_set(&pdu, AVTP_AAF_FIELD_SV, 1);
+    res = avtp_cvf_pdu_set(&pdu, AVTP_CVF_FIELD_SV, 1);
 
     assert_int_equal(res, 0);
     assert_true(ntohl(pdu.subtype_data) == 0x00800000);
@@ -310,12 +286,12 @@ static void aaf_set_field_sv(void **state)
     assert_true(pdu.packet_info == 0);
 }
 
-static void aaf_set_field_mr(void **state)
+static void cvf_set_field_mr(void **state)
 {
     int res;
     struct avtp_stream_pdu pdu = { 0 };
 
-    res = avtp_aaf_pdu_set(&pdu, AVTP_AAF_FIELD_MR, 1);
+    res = avtp_cvf_pdu_set(&pdu, AVTP_CVF_FIELD_MR, 1);
 
     assert_int_equal(res, 0);
     assert_true(ntohl(pdu.subtype_data) == 0x00080000);
@@ -325,12 +301,12 @@ static void aaf_set_field_mr(void **state)
     assert_true(pdu.packet_info == 0);
 }
 
-static void aaf_set_field_tv(void **state)
+static void cvf_set_field_tv(void **state)
 {
     int res;
     struct avtp_stream_pdu pdu = { 0 };
 
-    res = avtp_aaf_pdu_set(&pdu, AVTP_AAF_FIELD_TV, 1);
+    res = avtp_cvf_pdu_set(&pdu, AVTP_CVF_FIELD_TV, 1);
 
     assert_int_equal(res, 0);
     assert_true(ntohl(pdu.subtype_data) == 0x00010000);
@@ -340,12 +316,12 @@ static void aaf_set_field_tv(void **state)
     assert_true(pdu.packet_info == 0);
 }
 
-static void aaf_set_field_seq_num(void **state)
+static void cvf_set_field_seq_num(void **state)
 {
     int res;
     struct avtp_stream_pdu pdu = { 0 };
 
-    res = avtp_aaf_pdu_set(&pdu, AVTP_AAF_FIELD_SEQ_NUM, 0x55);
+    res = avtp_cvf_pdu_set(&pdu, AVTP_CVF_FIELD_SEQUENCE_NUM, 0x55);
 
     assert_int_equal(res, 0);
     assert_true(ntohl(pdu.subtype_data) == 0x00005500);
@@ -355,12 +331,12 @@ static void aaf_set_field_seq_num(void **state)
     assert_true(pdu.packet_info == 0);
 }
 
-static void aaf_set_field_tu(void **state)
+static void cvf_set_field_tu(void **state)
 {
     int res;
     struct avtp_stream_pdu pdu = { 0 };
 
-    res = avtp_aaf_pdu_set(&pdu, AVTP_AAF_FIELD_TU, 1);
+    res = avtp_cvf_pdu_set(&pdu, AVTP_CVF_FIELD_TU, 1);
 
     assert_int_equal(res, 0);
     assert_true(ntohl(pdu.subtype_data) == 0x00000001);
@@ -370,12 +346,12 @@ static void aaf_set_field_tu(void **state)
     assert_true(pdu.packet_info == 0);
 }
 
-static void aaf_set_field_stream_id(void **state)
+static void cvf_set_field_stream_id(void **state)
 {
     int res;
     struct avtp_stream_pdu pdu = { 0 };
 
-    res = avtp_aaf_pdu_set(&pdu, AVTP_AAF_FIELD_STREAM_ID,
+    res = avtp_cvf_pdu_set(&pdu, AVTP_CVF_FIELD_STREAM_ID,
                             0xAABBCCDDEEFF0001);
 
     assert_int_equal(res, 0);
@@ -386,12 +362,12 @@ static void aaf_set_field_stream_id(void **state)
     assert_true(pdu.packet_info == 0);
 }
 
-static void aaf_set_field_timestamp(void **state)
+static void cvf_set_field_timestamp(void **state)
 {
     int res;
     struct avtp_stream_pdu pdu = { 0 };
 
-    res = avtp_aaf_pdu_set(&pdu, AVTP_AAF_FIELD_TIMESTAMP, 0x80C0FFEE);
+    res = avtp_cvf_pdu_set(&pdu, AVTP_CVF_FIELD_AVTP_TIMESTAMP, 0x80C0FFEE);
 
     assert_int_equal(res, 0);
     assert_true(ntohl(pdu.avtp_time) == 0x80C0FFEE);
@@ -401,74 +377,44 @@ static void aaf_set_field_timestamp(void **state)
     assert_true(pdu.packet_info == 0);
 }
 
-static void aaf_set_field_format(void **state)
+static void cvf_set_field_format(void **state)
 {
     int res;
     struct avtp_stream_pdu pdu = { 0 };
 
-    res = avtp_aaf_pdu_set(&pdu, AVTP_AAF_FIELD_FORMAT,
-                        AVTP_AAF_FORMAT_INT_16BIT);
+    res = avtp_cvf_pdu_set(&pdu, AVTP_CVF_FIELD_FORMAT,
+                        AVTP_CVF_FORMAT_RFC);
 
     assert_int_equal(res, 0);
-    assert_true(ntohl(pdu.format_specific) == 0x04000000);
+    assert_true(ntohl(pdu.format_specific) == 0x02000000);
     assert_true(pdu.subtype_data == 0);
     assert_true(pdu.stream_id == 0);
     assert_true(pdu.avtp_time == 0);
     assert_true(pdu.packet_info == 0);
 }
 
-static void aaf_set_field_nsr(void **state)
+static void cvf_set_field_format_subtype(void **state)
 {
     int res;
     struct avtp_stream_pdu pdu = { 0 };
 
-    res = avtp_aaf_pdu_set(&pdu, AVTP_AAF_FIELD_NSR,
-                        AVTP_AAF_PCM_NSR_48KHZ);
+    res = avtp_cvf_pdu_set(&pdu, AVTP_CVF_FIELD_FORMAT_SUBTYPE,
+                        AVTP_CVF_FORMAT_SUBTYPE_H264);
 
     assert_int_equal(res, 0);
-    assert_true(ntohl(pdu.format_specific) == 0x00500000);
+    assert_true(ntohl(pdu.format_specific) == 0x10000);
     assert_true(pdu.subtype_data == 0);
     assert_true(pdu.stream_id == 0);
     assert_true(pdu.avtp_time == 0);
     assert_true(pdu.packet_info == 0);
 }
 
-static void aaf_set_field_chan(void **state)
+static void cvf_set_field_data_len(void **state)
 {
     int res;
     struct avtp_stream_pdu pdu = { 0 };
 
-    res = avtp_aaf_pdu_set(&pdu, AVTP_AAF_FIELD_CHAN_PER_FRAME, 0x2AA);
-
-    assert_int_equal(res, 0);
-    assert_true(ntohl(pdu.format_specific) == 0x0002AA00);
-    assert_true(pdu.subtype_data == 0);
-    assert_true(pdu.stream_id == 0);
-    assert_true(pdu.avtp_time == 0);
-    assert_true(pdu.packet_info == 0);
-}
-
-static void aaf_set_field_depth(void **state)
-{
-    int res;
-    struct avtp_stream_pdu pdu = { 0 };
-
-    res = avtp_aaf_pdu_set(&pdu, AVTP_AAF_FIELD_BIT_DEPTH, 0xA5);
-
-    assert_int_equal(res, 0);
-    assert_true(ntohl(pdu.format_specific) == 0x000000A5);
-    assert_true(pdu.subtype_data == 0);
-    assert_true(pdu.stream_id == 0);
-    assert_true(pdu.avtp_time == 0);
-    assert_true(pdu.packet_info == 0);
-}
-
-static void aaf_set_field_data_len(void **state)
-{
-    int res;
-    struct avtp_stream_pdu pdu = { 0 };
-
-    res = avtp_aaf_pdu_set(&pdu, AVTP_AAF_FIELD_STREAM_DATA_LEN, 0xAAAA);
+    res = avtp_cvf_pdu_set(&pdu, AVTP_CVF_FIELD_STREAM_DATA_LENGTH, 0xAAAA);
 
     assert_int_equal(res, 0);
     assert_true(ntohl(pdu.packet_info) == 0xAAAA0000);
@@ -478,12 +424,12 @@ static void aaf_set_field_data_len(void **state)
     assert_true(pdu.format_specific == 0);
 }
 
-static void aaf_set_field_sp(void **state)
+static void cvf_set_field_m(void **state)
 {
     int res;
     struct avtp_stream_pdu pdu = { 0 };
 
-    res = avtp_aaf_pdu_set(&pdu, AVTP_AAF_FIELD_SP, 1);
+    res = avtp_cvf_pdu_set(&pdu, AVTP_CVF_FIELD_M, 1);
 
     assert_int_equal(res, 0);
     assert_true(ntohl(pdu.packet_info) == 0x00001000);
@@ -493,12 +439,12 @@ static void aaf_set_field_sp(void **state)
     assert_true(pdu.format_specific == 0);
 }
 
-static void aaf_set_field_evt(void **state)
+static void cvf_set_field_evt(void **state)
 {
     int res;
     struct avtp_stream_pdu pdu = { 0 };
 
-    res = avtp_aaf_pdu_set(&pdu, AVTP_AAF_FIELD_EVT, 0xA);
+    res = avtp_cvf_pdu_set(&pdu, AVTP_CVF_FIELD_EVT, 0xA);
 
     assert_int_equal(res, 0);
     assert_true(ntohl(pdu.packet_info) == 0x00000A00);
@@ -508,68 +454,120 @@ static void aaf_set_field_evt(void **state)
     assert_true(pdu.format_specific == 0);
 }
 
-static void aaf_pdu_init_null_pdu(void **state)
+static void cvf_pdu_init_null_pdu(void **state)
 {
     int res;
 
-    res = avtp_aaf_pdu_init(NULL);
-
+    res = avtp_cvf_pdu_init(NULL, AVTP_CVF_FORMAT_SUBTYPE_H264);
     assert_int_equal(res, -EINVAL);
 }
 
-static void aaf_pdu_init(void **state)
+static void cvf_pdu_init(void **state)
 {
     int res;
     struct avtp_stream_pdu pdu;
 
-    res = avtp_aaf_pdu_init(&pdu);
+    res = avtp_cvf_pdu_init(&pdu, AVTP_CVF_FORMAT_SUBTYPE_H264);
 
     assert_int_equal(res, 0);
-    assert_true(ntohl(pdu.subtype_data) == 0x02800000);
+    assert_true(ntohl(pdu.subtype_data) == 0x03800000);
+    assert_true(pdu.stream_id == 0);
+    assert_true(pdu.avtp_time == 0);
+    assert_true(ntohl(pdu.format_specific) == 0x2010000);
+    assert_true(pdu.packet_info == 0);
+}
+
+static void cvf_get_field_ptv(void **state)
+{
+    int res;
+    uint64_t val;
+    struct avtp_stream_pdu pdu = { 0 };
+
+    /* Set 'ptv' field to 1. */
+    pdu.packet_info = htonl(0x00002000);
+
+    res = avtp_cvf_pdu_get(&pdu, AVTP_CVF_FIELD_PTV, &val);
+
+    assert_int_equal(res, 0);
+    assert_true(val == 1);
+}
+
+static void cvf_set_field_ptv(void **state)
+{
+    int res;
+    struct avtp_stream_pdu pdu = { 0 };
+
+    res = avtp_cvf_pdu_set(&pdu, AVTP_CVF_FIELD_PTV, 1);
+
+    assert_int_equal(res, 0);
+    assert_true(pdu.subtype_data == 0);
     assert_true(pdu.stream_id == 0);
     assert_true(pdu.avtp_time == 0);
     assert_true(pdu.format_specific == 0);
-    assert_true(pdu.packet_info == 0);
+    assert_true(ntohl(pdu.packet_info) == 0x00002000);
+}
+
+/**** Tests for H.264 fields ****/
+
+static void cvf_get_field_h264_timestamp(void **state)
+{
+    uint64_t val;
+    Avtp_H264_t pdu;
+
+    /* Set 'h264_timestamp' field (which lives in h264_header) to
+     * 0x80C0FFEE. */
+    uint32_t value = htonl(0x80C0FFEE);
+    memcpy(&pdu.header, &value, 4);
+
+    val = Avtp_H264_GetField(&pdu, AVTP_H264_FIELD_TIMESTAMP);
+    assert_true(val == 0x80C0FFEE);
+}
+
+static void cvf_set_field_h264_timestamp(void **state)
+{
+    Avtp_H264_t pdu;
+    Avtp_H264_SetField(&pdu, AVTP_H264_FIELD_TIMESTAMP, 0x80C0FFEE);
+    assert_true(ntohl(*(uint32_t*)(&pdu.header)) == 0x80C0FFEE);
 }
 
 int main(void)
 {
     const struct CMUnitTest tests[] = {
-        cmocka_unit_test(aaf_get_field_null_pdu),
-        cmocka_unit_test(aaf_get_field_null_val),
-        cmocka_unit_test(aaf_get_field_invalid_field),
-        cmocka_unit_test(aaf_get_field_sv),
-        cmocka_unit_test(aaf_get_field_mr),
-        cmocka_unit_test(aaf_get_field_tv),
-        cmocka_unit_test(aaf_get_field_seq_num),
-        cmocka_unit_test(aaf_get_field_tu),
-        cmocka_unit_test(aaf_get_field_stream_id),
-        cmocka_unit_test(aaf_get_field_timestamp),
-        cmocka_unit_test(aaf_get_field_format),
-        cmocka_unit_test(aaf_get_field_nsr),
-        cmocka_unit_test(aaf_get_field_chan),
-        cmocka_unit_test(aaf_get_field_depth),
-        cmocka_unit_test(aaf_get_field_data_len),
-        cmocka_unit_test(aaf_get_field_sp),
-        cmocka_unit_test(aaf_get_field_evt),
-        cmocka_unit_test(aaf_set_field_null_pdu),
-        cmocka_unit_test(aaf_set_field_invalid_field),
-        cmocka_unit_test(aaf_set_field_sv),
-        cmocka_unit_test(aaf_set_field_mr),
-        cmocka_unit_test(aaf_set_field_tv),
-        cmocka_unit_test(aaf_set_field_seq_num),
-        cmocka_unit_test(aaf_set_field_tu),
-        cmocka_unit_test(aaf_set_field_stream_id),
-        cmocka_unit_test(aaf_set_field_timestamp),
-        cmocka_unit_test(aaf_set_field_format),
-        cmocka_unit_test(aaf_set_field_nsr),
-        cmocka_unit_test(aaf_set_field_chan),
-        cmocka_unit_test(aaf_set_field_depth),
-        cmocka_unit_test(aaf_set_field_data_len),
-        cmocka_unit_test(aaf_set_field_sp),
-        cmocka_unit_test(aaf_set_field_evt),
-        cmocka_unit_test(aaf_pdu_init_null_pdu),
-        cmocka_unit_test(aaf_pdu_init),
+        cmocka_unit_test(cvf_get_field_null_pdu),
+        cmocka_unit_test(cvf_get_field_null_val),
+        cmocka_unit_test(cvf_get_field_invalid_field),
+        cmocka_unit_test(cvf_get_field_sv),
+        cmocka_unit_test(cvf_get_field_mr),
+        cmocka_unit_test(cvf_get_field_tv),
+        cmocka_unit_test(cvf_get_field_seq_num),
+        cmocka_unit_test(cvf_get_field_tu),
+        cmocka_unit_test(cvf_get_field_stream_id),
+        cmocka_unit_test(cvf_get_field_timestamp),
+        cmocka_unit_test(cvf_get_field_format),
+        cmocka_unit_test(cvf_get_field_format_subtype),
+        cmocka_unit_test(cvf_get_field_data_len),
+        cmocka_unit_test(cvf_get_field_m),
+        cmocka_unit_test(cvf_get_field_evt),
+        cmocka_unit_test(cvf_get_field_ptv),
+        cmocka_unit_test(cvf_get_field_h264_timestamp),
+        cmocka_unit_test(cvf_set_field_null_pdu),
+        cmocka_unit_test(cvf_set_field_invalid_field),
+        cmocka_unit_test(cvf_set_field_sv),
+        cmocka_unit_test(cvf_set_field_mr),
+        cmocka_unit_test(cvf_set_field_tv),
+        cmocka_unit_test(cvf_set_field_seq_num),
+        cmocka_unit_test(cvf_set_field_tu),
+        cmocka_unit_test(cvf_set_field_stream_id),
+        cmocka_unit_test(cvf_set_field_timestamp),
+        cmocka_unit_test(cvf_set_field_format),
+        cmocka_unit_test(cvf_set_field_format_subtype),
+        cmocka_unit_test(cvf_set_field_data_len),
+        cmocka_unit_test(cvf_set_field_m),
+        cmocka_unit_test(cvf_set_field_evt),
+        cmocka_unit_test(cvf_set_field_ptv),
+        cmocka_unit_test(cvf_set_field_h264_timestamp),
+        cmocka_unit_test(cvf_pdu_init_null_pdu),
+        cmocka_unit_test(cvf_pdu_init),
     };
 
     return cmocka_run_group_tests(tests, NULL, NULL);
