+++ unit/test-gpc.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2025, COVESA
+ * Copyright (c) 2024, COVESA
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
@@ -30,66 +30,67 @@
 #include <stdarg.h>
 #include <stddef.h>
 #include <setjmp.h>
+extern "C" {
 #include <cmocka.h>
+}
 #include <arpa/inet.h>
-#include <errno.h>
 #include <string.h>
-#include <math.h>
-#include <stdio.h>
 
-#include "avtp/acf/Tscf.h"
-#include "avtp/CommonHeader.h"
+#include "avtp/acf/Gpc.h"
 
-#define MAX_PDU_SIZE        1500
+#define MAX_PDU_SIZE 1500
 
-static void tscf_init(void **state) {
-
+static void gpc_init(void **state) {
     uint8_t pdu[MAX_PDU_SIZE];
-    uint8_t init_pdu[AVTP_TSCF_HEADER_LEN];
+    uint8_t init_pdu[AVTP_GPC_HEADER_LEN];
 
-    // Check the lengths
-    assert_int_equal(sizeof(Avtp_Tscf_t), AVTP_TSCF_HEADER_LEN);
+    Avtp_Gpc_Init(NULL);
 
-    // Check init function while passing in a null pointer
-    Avtp_Tscf_Init(NULL);
-
-    // Check if the function is initializing properly
-    Avtp_Tscf_Init((Avtp_Tscf_t*)pdu);
-    memset(init_pdu, 0, AVTP_TSCF_HEADER_LEN);
-    init_pdu[0] = AVTP_SUBTYPE_TSCF; // Setting AVTP Subtype as TSCF
-    init_pdu[1] = 0x80; // Setting Stream as valid
-    assert_memory_equal(init_pdu, pdu, AVTP_TSCF_HEADER_LEN);
+    Avtp_Gpc_Init((Avtp_Gpc_t*)pdu);
+    memset(init_pdu, 0, AVTP_GPC_HEADER_LEN);
+    init_pdu[0] = AVTP_ACF_TYPE_GPC << 1;
+    assert_memory_equal(init_pdu, pdu, AVTP_GPC_HEADER_LEN);
 }
 
-static void tscf_is_valid(void **state) {
+static void gpc_get_set_fields(void **state) {
+    uint8_t pdu[MAX_PDU_SIZE];
+    
+    Avtp_Gpc_Init((Avtp_Gpc_t*)pdu);
 
-    uint8_t pdu[MAX_PDU_SIZE], result;
+    Avtp_Gpc_SetAcfMsgType((Avtp_Gpc_t*)pdu, AVTP_ACF_TYPE_GPC);
+    assert_int_equal(Avtp_Gpc_GetAcfMsgType((Avtp_Gpc_t*)pdu), AVTP_ACF_TYPE_GPC);
 
-    // Valid IEEE 1722 TSCF Frame
-    Avtp_Tscf_Init((Avtp_Tscf_t*)pdu);
-    assert_int_equal(Avtp_Tscf_IsValid((Avtp_Tscf_t*)pdu, MAX_PDU_SIZE), 1);
+    Avtp_Gpc_SetAcfMsgLength((Avtp_Gpc_t*)pdu, 200);
+    assert_int_equal(Avtp_Gpc_GetAcfMsgLength((Avtp_Gpc_t*)pdu), 200);
 
-    // Not a IEEE 1722 TSCF Frame
+    Avtp_Gpc_SetGpcMsgId((Avtp_Gpc_t*)pdu, 0x456789ABCDEFULL);
+    assert_int_equal(Avtp_Gpc_GetGpcMsgId((Avtp_Gpc_t*)pdu), 0x456789ABCDEFULL);
+}
+
+static void gpc_is_valid(void **state) {
+    uint8_t pdu[MAX_PDU_SIZE];
+
+    Avtp_Gpc_Init((Avtp_Gpc_t*)pdu);
+    assert_int_equal(Avtp_Gpc_IsValid((Avtp_Gpc_t*)pdu, MAX_PDU_SIZE), 1);
+
     memset(pdu, 0, MAX_PDU_SIZE);
-    assert_int_equal(Avtp_Tscf_IsValid((Avtp_Tscf_t*)pdu, MAX_PDU_SIZE), 0);
+    assert_int_equal(Avtp_Gpc_IsValid((Avtp_Gpc_t*)pdu, MAX_PDU_SIZE), 0);
 
-    // Valid IEEE 1722 TSCF Frame (Length 28, Buffer 30)
-    Avtp_Tscf_Init((Avtp_Tscf_t*)pdu);
-    Avtp_Tscf_SetStreamDataLength((Avtp_Tscf_t*)pdu, 28);
-    assert_int_equal(Avtp_Tscf_IsValid((Avtp_Tscf_t*)pdu, 30), 1);
+    Avtp_Gpc_Init((Avtp_Gpc_t*)pdu);
+    Avtp_Gpc_SetAcfMsgLength((Avtp_Gpc_t*)pdu, 4);
+    assert_int_equal(Avtp_Gpc_IsValid((Avtp_Gpc_t*)pdu, 20), 1);
 
-    // Invalid IEEE 1722 TSCF Frame (Length 24 but buffer only 9!)
-    Avtp_Tscf_Init((Avtp_Tscf_t*)pdu);
-    Avtp_Tscf_SetStreamDataLength((Avtp_Tscf_t*)pdu, 24);
-    assert_int_equal(Avtp_Tscf_IsValid((Avtp_Tscf_t*)pdu, 9), 0);
-
+    Avtp_Gpc_Init((Avtp_Gpc_t*)pdu);
+    Avtp_Gpc_SetAcfMsgLength((Avtp_Gpc_t*)pdu, 4);
+    assert_int_equal(Avtp_Gpc_IsValid((Avtp_Gpc_t*)pdu, 5), 0);
 }
 
 int main(void)
 {
     const struct CMUnitTest tests[] = {
-        cmocka_unit_test(tscf_init),
-        cmocka_unit_test(tscf_is_valid),
+        cmocka_unit_test(gpc_init),
+        cmocka_unit_test(gpc_get_set_fields),
+        cmocka_unit_test(gpc_is_valid),
     };
 
     return cmocka_run_group_tests(tests, NULL, NULL);
