+++ unit/test-ntscf.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2024, COVESA
+ * Copyright (c) 2025, COVESA
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
@@ -30,65 +30,68 @@
 #include <stdarg.h>
 #include <stddef.h>
 #include <setjmp.h>
+extern "C" {
 #include <cmocka.h>
+}
 #include <arpa/inet.h>
+#include <errno.h>
 #include <string.h>
+#include <math.h>
+#include <stdio.h>
 
-#include "avtp/acf/Gpc.h"
+#include "avtp/acf/Ntscf.h"
+#include "avtp/CommonHeader.h"
 
-#define MAX_PDU_SIZE 1500
+#define MAX_PDU_SIZE        1500
 
-static void gpc_init(void **state) {
+static void ntscf_init(void **state) {
+
     uint8_t pdu[MAX_PDU_SIZE];
-    uint8_t init_pdu[AVTP_GPC_HEADER_LEN];
+    uint8_t init_pdu[AVTP_NTSCF_HEADER_LEN];
 
-    Avtp_Gpc_Init(NULL);
+    // Check the lengths
+    assert_int_equal(sizeof(Avtp_Ntscf_t), AVTP_NTSCF_HEADER_LEN);
 
-    Avtp_Gpc_Init((Avtp_Gpc_t*)pdu);
-    memset(init_pdu, 0, AVTP_GPC_HEADER_LEN);
-    init_pdu[0] = AVTP_ACF_TYPE_GPC << 1;
-    assert_memory_equal(init_pdu, pdu, AVTP_GPC_HEADER_LEN);
+    // Check init function while passing in a null pointer
+    Avtp_Ntscf_Init(NULL);
+
+    // Check if the function is initializing properly
+    Avtp_Ntscf_Init((Avtp_Ntscf_t*)pdu);
+    memset(init_pdu, 0, AVTP_NTSCF_HEADER_LEN);
+    init_pdu[0] = AVTP_SUBTYPE_NTSCF; // Setting AVTP Subtype as NTSCF
+    init_pdu[1] = 0x80; // Setting Stream as valid
+    assert_memory_equal(init_pdu, pdu, AVTP_NTSCF_HEADER_LEN);
 }
 
-static void gpc_get_set_fields(void **state) {
-    uint8_t pdu[MAX_PDU_SIZE];
-    
-    Avtp_Gpc_Init((Avtp_Gpc_t*)pdu);
+static void ntscf_is_valid(void **state) {
 
-    Avtp_Gpc_SetAcfMsgType((Avtp_Gpc_t*)pdu, AVTP_ACF_TYPE_GPC);
-    assert_int_equal(Avtp_Gpc_GetAcfMsgType((Avtp_Gpc_t*)pdu), AVTP_ACF_TYPE_GPC);
+    uint8_t pdu[MAX_PDU_SIZE], result;
 
-    Avtp_Gpc_SetAcfMsgLength((Avtp_Gpc_t*)pdu, 200);
-    assert_int_equal(Avtp_Gpc_GetAcfMsgLength((Avtp_Gpc_t*)pdu), 200);
+    // Valid IEEE 1722 NTSCF Frame
+    Avtp_Ntscf_Init((Avtp_Ntscf_t*)pdu);
+    assert_int_equal(Avtp_Ntscf_IsValid((Avtp_Ntscf_t*)pdu, MAX_PDU_SIZE), 1);
 
-    Avtp_Gpc_SetGpcMsgId((Avtp_Gpc_t*)pdu, 0x456789ABCDEFULL);
-    assert_int_equal(Avtp_Gpc_GetGpcMsgId((Avtp_Gpc_t*)pdu), 0x456789ABCDEFULL);
-}
-
-static void gpc_is_valid(void **state) {
-    uint8_t pdu[MAX_PDU_SIZE];
-
-    Avtp_Gpc_Init((Avtp_Gpc_t*)pdu);
-    assert_int_equal(Avtp_Gpc_IsValid((Avtp_Gpc_t*)pdu, MAX_PDU_SIZE), 1);
-
+    // Not a IEEE 1722 NTSCF Frame
     memset(pdu, 0, MAX_PDU_SIZE);
-    assert_int_equal(Avtp_Gpc_IsValid((Avtp_Gpc_t*)pdu, MAX_PDU_SIZE), 0);
+    assert_int_equal(Avtp_Ntscf_IsValid((Avtp_Ntscf_t*)pdu, MAX_PDU_SIZE), 0);
 
-    Avtp_Gpc_Init((Avtp_Gpc_t*)pdu);
-    Avtp_Gpc_SetAcfMsgLength((Avtp_Gpc_t*)pdu, 4);
-    assert_int_equal(Avtp_Gpc_IsValid((Avtp_Gpc_t*)pdu, 20), 1);
+    // Valid IEEE 1722 NTSCF Frame (Length 28, Buffer 30)
+    Avtp_Ntscf_Init((Avtp_Ntscf_t*)pdu);
+    Avtp_Ntscf_SetNtscfDataLength((Avtp_Ntscf_t*)pdu, 28);
+    assert_int_equal(Avtp_Ntscf_IsValid((Avtp_Ntscf_t*)pdu, 30), 1);
 
-    Avtp_Gpc_Init((Avtp_Gpc_t*)pdu);
-    Avtp_Gpc_SetAcfMsgLength((Avtp_Gpc_t*)pdu, 4);
-    assert_int_equal(Avtp_Gpc_IsValid((Avtp_Gpc_t*)pdu, 5), 0);
+    // Invalid IEEE 1722 NTSCF Frame (Length 24 but buffer only 9!)
+    Avtp_Ntscf_Init((Avtp_Ntscf_t*)pdu);
+    Avtp_Ntscf_SetNtscfDataLength((Avtp_Ntscf_t*)pdu, 24);
+    assert_int_equal(Avtp_Ntscf_IsValid((Avtp_Ntscf_t*)pdu, 9), 0);
+
 }
 
 int main(void)
 {
     const struct CMUnitTest tests[] = {
-        cmocka_unit_test(gpc_init),
-        cmocka_unit_test(gpc_get_set_fields),
-        cmocka_unit_test(gpc_is_valid),
+        cmocka_unit_test(ntscf_init),
+        cmocka_unit_test(ntscf_is_valid)
     };
 
     return cmocka_run_group_tests(tests, NULL, NULL);
